Creacion de modulo registrar

+++ database/migrations/2023_09_27_140837_crete_ubicaciones_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('ubicaciones', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('ruta_id');
+            $table->unsignedBigInteger('ruta_id');
             $table->string('posX');
             $table->string('posY');
             $table->string('nombre');
